Build number: 8908313405312549568 (latest to pass QA)
Release tag: 20190710_01_RC00

+++ src/connectivity/overnet/lib/embedded/decoder.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <fuchsia/overnet/protocol/cpp/fidl.h>
+
 #include "src/connectivity/overnet/lib/endpoint/router_endpoint.h"
 #include "src/connectivity/overnet/lib/vocabulary/optional.h"
 
@@ -56,7 +57,7 @@ class Decoder {
     return *reinterpret_cast<const fidl_message_header_t*>(
         message_.bytes.data());
   }
-  uint32_t ordinal() const { return header().ordinal; }
+  uint64_t ordinal() const { return header().ordinal; }
 
  private:
   fuchsia::overnet::protocol::ZirconChannelMessage message_;
